database_ingestion/app/requirements.txt
Flask
Flask-Cors
pandas
requests
sqlalchemy
mysqlclient

